cleanup: move file into common folder
